Worker: Rename worker components to clarify what they work for
(1) It sounds like WorkerGlobalScopeProxy is the base class of
ServiceWorkerGlobalScopeProxy, but it is not true. WorkerGlobalScopeProxy works
for in-process workers, namely, DedicatedWorker and CompositorWorker.
This patch renames them as follows:

  WorkerGlobalScopeProxy => InProcessWorkerGlobalScopeProxy
  WorkerMessagingProxy => InProcessWorkerMessagingProxy

(2) WorkerGlobalScopeProxyProvider provides a proxy for DedicatedWorker, so it
should be prefixed with "Dedicated":

  WorkerGlobalScopeProxyProvider => DedicatedWorkerGlobalScopeProxyProvider
  WorkerGlobalScopeProxyProviderImpl => DedicatedWorkerGlobalScopeProxyProviderImpl

BUG=603785

Review URL: https://codereview.chromium.org/1888703002

Cr-Original-Commit-Position: refs/heads/master@{#388166}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c4334f65977df7d8c44f14b7cf2ecc2f32772a2a