Notify Blink about start of gesture scroll through a queued event.
Added WebIputEventL::TouchScrollStarted event to notify Blink when a
gesture scroll is starting.

We need to know from Blink when gesture scroll is about to start.
So far, the only use of this info is firing a pointercancel event
during scrolling. We had been using a bit in WebTouchEvent before,
which didn't really reflect the possibility of scrolling (only
indicates if the touch has moved beyond the slop region). To find
the correct info to pass to Blink, we have to consider the effect of
touch-action on gestures.

In this CL, we push a new event into the touch queue when gestures
passes touch-action filtering.

This CL also adds a hack to prevent unintended reentry to a dispatch
method in TouchEventQueue.

BUG=567740,600773

Review URL: https://codereview.chromium.org/1800143002

Cr-Original-Commit-Position: refs/heads/master@{#388258}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b3c44a0beb7e5a56fc78861c22b588380de3f49b