Refactor SkiaImageFilterBuilder into a namespace.
This object has no state, no constructor logic, and an empty (albeit outlined)
destructor. Groups of this sort should be, at most, a namespace. This tears
out all of the plumbing of a reference to this object, when we can just refer
to it statically.

Review URL: https://codereview.chromium.org/1897333002

Cr-Original-Commit-Position: refs/heads/master@{#388281}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 26a7c61ac75372f18bd45fa59a643015759028d7