Don't restart the ImageQualityController timer unless it is already half over.
Restarting the timer every time an image continues to animate is very expensive,
because there is no way to actuually cancel the existing timer. Instead, the timer
is neutered, so that when it fires it will do nothing. This results in a flood
of no-op timer callbacks in the referenced bug scenario.

Instead, only restart the timer if the current frame time is more than half
of the timer time later than the frame time at the start of the timer.
To accomplish that, record the current frame time in WebViewImpl, so that it
can be compared with a future frame time.

We do this plumbing to avoid making lots of calls to query the system time,
which is expensive. The compositor has a similar optimization; see
https://code.google.com/p/chromium/codesearch#chromium/src/cc/layers/picture_layer_impl.cc&sq=package:chromium&l=452
for example.

BUG=605701

Review URL: https://codereview.chromium.org/1911273002

Cr-Original-Commit-Position: refs/heads/master@{#389394}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f79f4d6a1c52efbe06e30ace1ce882ac1d53aa40