Add WebUSB layout tests that detach and navigate an iframe.
The device should automatically be disconnected when the page is closed
or the frame navigates away. These tests create an iframe in order to
test that. An equivalent to Binding::set_connection_error_handler had
to be added to the Javascript bindings.

BUG=None

Review URL: https://codereview.chromium.org/1915703002

Cr-Original-Commit-Position: refs/heads/master@{#389541}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8e812ccc291940bef4f955ce34d15aa99169efe9