QuotaTracker: Use 'raw' string as key for origin
Prior to r382660 the QuotaTracker API took DatabaseIdentifiers, and a
serialization was used as the key in maps. This was changed to take
SecurityOrigins instead and the serialization via toString(). This
introduced additional special cases e.g. "null" for file:// origins in
some cases, leading to mismatches between setting and lookup in the
map depending on when the SecurityOrigin was minted. To fix, use
toRawString() which matches the old behavior without special cases.

BUG=604811
R=michaeln@chromium.org

Review URL: https://codereview.chromium.org/1914763002

Cr-Original-Commit-Position: refs/heads/master@{#389607}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 976f8fa83bf90a5983af3ca8a6e74d34044b93ff