Get the right ascent/descent for the tallest vertically aligned container
When we're figuring out the right adjustment to make to the ascent/descent in
a line to fill the full height of the line with vertical align top/bottom we
shouldn't carry forward the results of previous speculative adjustments from
one box on the line to the next.

Also ascent/descent are round()'ed so do the same to the lineHeight we're fitting
to.

BUG=604144

Review URL: https://codereview.chromium.org/1910053003

Cr-Original-Commit-Position: refs/heads/master@{#389668}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 020751ff9d9524fe83cc858edf4d1aea78e305ae