Prevent markContainerChainForLayout to scheduleRelayout() if isInPerformLayout()
When we're in layout, markContainerChainForLayout() is marking a
descendant as needing layout with the intention of visiting it during
this layout. We shouldn't be scheduling it to be laid out later.

Also, schedduleRelayout() must not be called while iterating
FrameView::m_layoutSubtreeRootList.

BUG=602908, 589773

Review URL: https://codereview.chromium.org/1906803002

Cr-Original-Commit-Position: refs/heads/master@{#389691}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1a0aa2dff16abbd9bc7f906b4d807766668bde17