Add XML header to test SVG file
WorldIso.svg is the only file in the SVG performance test suite
that's missing the xml header. This triggered auto-encoding detector or
UTF-8 encoding detector and caused extra amount of time in the result,
which is not the purpose of the test suite. Added a header used
in other svg (Worldcup.svg). This fixes the regression reported
in blink_perf.svg.

BUG=606267

Review URL: https://codereview.chromium.org/1922843002

Cr-Original-Commit-Position: refs/heads/master@{#389923}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 472f4e7c9b1ce63e453be41e938520dec30af46b