Do not call RenderViewObserver::Navigate for subframes
This CL makes it so that RenderFrameImpl::PrepareRenderViewForNavigation will
not call RenderViewObserver::Navigate when RenderFrameImpl is a subframe. This
is causing an issue with PlzNavigate in layout tests, where the BlinkTestRunner
would cause the subframe to be focused on commit, which is unexpected.

BUG=576261

Review-Url: https://codereview.chromium.org/1921223004
Cr-Original-Commit-Position: refs/heads/master@{#390111}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d4c492c56cb8342ab2f2794a10ec9a10ed71cacb