Prevent sending click event for non primary button
Prevent sending click and dblclick events when the
button in charge is a non-primary mouse button.

As the result of this change there is no way to
prevent the behavior of opening a new tap when
middle clicking on a link. Previously, js could
listen to click event of the middle button and
call preventDefault on it.

BUG=255

Review-Url: https://codereview.chromium.org/1894253002
Cr-Original-Commit-Position: refs/heads/master@{#390195}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 76fea00a18f75886ea649414393228180306e13d