Implementing document.setRootScroller API for main thread scrolling.
This patch implements the setRootScroller API, also know as the non-document
root scroller proposal. See this expaliner for details:
https://github.com/bokand/NonDocumentRootScroller

This patch implements most of the initial functionality but only for scrolling
on the main thread. All new API surfaces are behind a RuntimeEnabledFeature.

BUG=505516

Review-Url: https://codereview.chromium.org/1913843004
Cr-Original-Commit-Position: refs/heads/master@{#390200}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 861d774435b04448856e99462d4f4e3d734f72f4