Test fragment scrolling and history restoration interaction w/ scroll anchoring.
This CL allows RuntimeEnabledFeatures with the option 'set_from_internals' to be set
from Internals. Previously, we could only read whether a feature was enabled.

The tests in this CL verify that we do scroll anchoring when a page has a fragment
and when doing history restoration.

Note that there are no behavioral changes in this CL.

BUG=594880,594879,537764

Review-Url: https://codereview.chromium.org/1895293002
Cr-Original-Commit-Position: refs/heads/master@{#390463}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5114bd75de8b0e7813df98a6a2c9b43943cd4229