Fix crash when getting main frame layout size creating FrameView.
It looks like webView->page() is null so when we try to get the
mainFrameLayoutSize, which requires getting the minimum scale, we try to ge
PageScaleConstraintsSet from frameHost via the null Page pointer. This seems
like a case where we've started shutting down, but then we get requests coming
back from the server and trying to create a FrameView.

This patch fixes the issue by returning without performing a NULL check before
creating the FrameView so we can early-out.

BUG=606735

Review-Url: https://codereview.chromium.org/1929873003
Cr-Original-Commit-Position: refs/heads/master@{#390529}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b01b27dee20f70ab3940aaeb582568804fa7e2f3