If the <use> is hidden and the child of it is visible, should clip the non-resource.
If <use> element that is child of <clipPath> element is hidden
and the child of it is visible, should clip the non-resource element.
Because check the child of <use> element whether is visible or not.
If it is visible, contains on clip boundaries.

There is the change one more on this test named
'clip-path-referencing-use2.svg' is "50x50" -> "0x0" change.
Because if the clipping target doesn't need clipping,
doesn't contain clip boundary about <use> element on
calculateClipContentPaintInvalidationRect function.

And remove reportError() to print error info for <clipPath> usage
on SVGUseElement.

BUG=590153

Review-Url: https://codereview.chromium.org/1839973003
Cr-Original-Commit-Position: refs/heads/master@{#390669}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 53cbed151cea16f4d27cc887a272d9e672c41b75