Provide shipping in PaymentRequest result
If the merchant provides non-NULL shipping options, the merchant does
not need shipping address to calculate shipping costs. The UI selects
the default shipping address, but does not send it to the merchant. This
patch sends the shipping address to the merchant when the user clicks
"Pay".

BUG=608222

Review-Url: https://codereview.chromium.org/1937783002
Cr-Original-Commit-Position: refs/heads/master@{#391133}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f63c3bce0928c9534671fb325d45fd1524ac6b37